apify/kranzer/storelocator/amctheatres_com/requirements.txt
bs4==0.0.1
requests==2.11.1
w3lib
sgvalidator==0.0.19
sgrequests
lxmlsglogging
